fix(jolt-instructions): execute/table consistency, compile-time assertions, and suffixes allocation
- Fix W-suffix execute()/table sign-extension mismatch: VirtualXorRotW{16,12,8,7}
  and VirtualRotriw were sign-extending results but their tables zero-extend.
  Aligned execute() to match tables.
- Fix VirtualChangeDivisorW execute() to explicitly sign-extend, matching table.
- Add compile-time const assertions for NUM_SUFFIXES, NUM_CIRCUIT_FLAGS,
  and NUM_INSTRUCTION_FLAGS (matching existing LookupTableKind::COUNT pattern).
- Change PrefixSuffixDecomposition::suffixes() to return &'static [Suffixes]
  instead of Vec<Suffixes>, avoiding heap allocation in prover hot path.
- Add debug_assert to LookupBits::pop_msb for empty bitvector.
- Remove unnecessary #![allow(unused_results)] from lookup_table_tests.
- Add WHY comments for shift/W-suffix instructions with flag/table mismatches.

+++ crates/jolt-instructions/src/virtual_/shift.rs
@@ -83,7 +83,7 @@ define_instruction!(
 define_instruction!(
     /// Virtual ROTRIW: 32-bit rotate right by immediate, sign-extended.
     VirtualRotriw, opcodes::VIRTUAL_ROTRIW, "VIRTUAL_ROTRIW",
-    |x, y| (x as u32).rotate_right((y & 31) as u32) as i32 as i64 as u64,
+    |x, y| (x as u32).rotate_right((y & 31) as u32) as u64,
     circuit: [WriteLookupOutputToRD],
     instruction: [LeftOperandIsRs1Value, RightOperandIsImm],
     table: VirtualROTRW,
